Updated for 7.1.0.22900

+++ SpellWork/Spell/SpellInfo.cs
@@ -580,7 +580,7 @@ namespace SpellWork.Spell
                             var gtScaling = GameTable<GtSpellScalingEntry>.GetRecord(level);
                             Debug.Assert(gtScaling != null);
                             value = Scaling.ScalingClass > 0
-                                ? gtScaling.GetColumnForClass(Scaling.ScalingClass)
+                                ? (float)gtScaling.GetColumnForClass(Scaling.ScalingClass)
                                 : gtScaling.Item;
                         }
                         else if (DBC.DBC.RandPropPoints.ContainsKey(level))
